Refactor Import History components for improved functionality and user experience
- Introduced new formatting methods in ImportHistoryDataPreview.vue for boolean, number, and URL display.
- Removed the ImportHistoryDetails component and its associated view model, streamlining the import history handling.
- Updated ImportHistoryList.vue to simplify the data loading process and removed unnecessary filters.
- Enhanced the index.vue to manage import history modals more effectively, improving user interaction.
- Adjusted the GetImportHistoryUseCase to limit filters to workspace_id and user_id for cleaner API requests.
