+++ argilla-frontend/pages/index.vue
@@ -104,6 +104,7 @@
       @close="showImportModal = false"
     />
 
+    <!-- Import History Modal -->
     <BaseModal
       :modal-visible="isImportHistoryModalVisible"
       @close-modal="closeImportHistoryModal"
@@ -116,6 +117,23 @@
         @close="closeImportHistoryModal"
       />
     </BaseModal>
+
+    <!-- Import History Details Modal -->
+    <BaseModal
+      :modal-visible="isImportDetailsModalVisible"
+      @close-modal="closeImportDetailsModal"
+      :modal-title="`Import Details - ${selectedImportDetails?.filename || 'Unknown'}`"
+      modal-class="modal-large"
+    >
+      <ImportHistoryDetailsModal
+        v-if="selectedImportDetails"
+        :import-id="selectedImportDetails.importId"
+        :filename="selectedImportDetails.filename"
+        :workspace="selectedImportDetails.workspace"
+        @close="closeImportDetailsModal"
+        @retry-item="handleRetryItem"
+      />
+    </BaseModal>
   </div>
 </template>
 
@@ -123,7 +141,7 @@
 import Home from "@/layouts/Home.vue";
 import { useHomeViewModel } from "./useHomeViewModel";
 import { Workspace } from "~/v1/domain/entities/workspace/Workspace";
-
+import ImportHistoryDetailsModal from "~/components/features/import/ImportHistoryDetailsModal.vue";
 
 export default {
   data() {
@@ -135,6 +153,9 @@ export default {
         { id: 'datasets', name: this.$t('home.datasets') },
         { id: 'documents', name: this.$t('home.documents') },
       ],
+      // Import details modal state
+      isImportDetailsModalVisible: false,
+      selectedImportDetails: null,
     };
   },
   methods: {
@@ -164,12 +185,21 @@ export default {
       this.goToImportConfiguration(importRecord.id);
     },
     handleViewImportDetails(importRecord) {
-      this.closeImportHistoryModal();
-      this.goToImportConfiguration(importRecord.id);
+      this.selectedImportDetails = importRecord;
+      this.isImportDetailsModalVisible = true;
+    },
+    closeImportDetailsModal() {
+      this.isImportDetailsModalVisible = false;
+      this.selectedImportDetails = null;
+    },
+    handleRetryItem(item) {
+      // Handle retry item functionality if needed
+      console.log('Retry item:', item);
     },
   },
   components: {
     Home,
+    ImportHistoryDetailsModal,
   },
   computed: {
     // Modal state is managed by useHomeViewModel
